Fixed routes so that announcement routing is in a separate class
